# web-rally/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Rally Tascas API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/rally/v1/activities/deferred/{activity_id}/capture": {
      "post": {
        "tags": [
          "Deferred Judging"
        ],
        "summary": "Capture Deferred Result",
        "operationId": "capture_deferred_result",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "activity_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Activity Id"
            }
          },
          {
            "name": "team_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Team Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_capture_deferred_result"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeferredResultResponse"
                }
              }
            }
          },
          "404": {
            "description": "Activity not found"
          },
          "400": {
            "description": "Activity is not deferred-judged type, or team_id is missing"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/rally/v1/activities/results/{result_id}/judge": {
      "put": {
        "tags": [
          "Deferred Judging"
        ],
        "summary": "Judge Deferred Result",
        "operationId": "judge_deferred_result",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "result_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Result Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JudgeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeferredResultResponse"
                }
              }
            }
          },
          "404": {
            "description": "Result not found"
          },
          "400": {
            "description": "Result is not pending judgment"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/rally/v1/activities/results/{result_id}/set-team-photo": {
      "put": {
        "tags": [
          "Deferred Judging"
        ],
        "summary": "Set Team Photo From Result",
        "description": "Promote one of a deferred-judging result's submitted photos to the team's official photo.\n\nGated by rally_settings.allow_photo_as_team_photo so an admin can turn\nthis capability off event-wide. The chosen URL must already be one of\nthe result's own media_urls (already stored in R2) to prevent staff\nfrom pointing a team's photo at an arbitrary URL.",
        "operationId": "set_team_photo_from_result",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "result_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Result Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetTeamPhotoRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SetTeamPhotoResponse"
                }
              }
            }
          },
          "403": {
            "description": "Setting a team photo from an activity photo is disabled"
          },
          "404": {
            "description": "Result not found"
          },
          "400": {
            "description": "Photo does not belong to this result"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/rally/v1/activities/deferred/pending": {
      "get": {
        "tags": [
          "Deferred Judging"
        ],
        "summary": "List Pending Judgments",
        "operationId": "list_pending_judgments",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DeferredResultResponse"
                  },
                  "type": "array",
                  "title": "Response List Pending Judgments"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_capture_deferred_result": {
        "properties": {
          "images": {
            "items": {
              "type": "string",
              "contentMediaType": "application/octet-stream"
            },
            "type": "array",
            "title": "Images"
          }
        },
        "type": "object",
        "title": "Body_capture_deferred_result"
      },
...
      "DeferredResultResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "activity_id": {
            "type": "integer",
            "title": "Activity Id"
          },
          "team_id": {
            "type": "integer",
            "title": "Team Id"
          },
          "judgment_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Judgment Status"
          },
          "media_urls": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Media Urls"
          },
          "final_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Final Score"
          },
          "is_completed": {
            "type": "boolean",
            "title": "Is Completed"
          }
        },
        "type": "object",
        "required": [
          "id",
          "activity_id",
          "team_id",
          "media_urls",
          "is_completed"
        ],
        "title": "DeferredResultResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "JudgeRequest": {
        "properties": {
          "points": {
            "type": "number",
            "title": "Points"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "type": "object",
        "required": [
          "points"
        ],
        "title": "JudgeRequest"
      },
...
      "SetTeamPhotoRequest": {
        "properties": {
          "image_url": {
            "type": "string",
            "title": "Image Url"
          }
        },
        "type": "object",
        "required": [
          "image_url"
        ],
        "title": "SetTeamPhotoRequest"
      },
      "SetTeamPhotoResponse": {
        "properties": {
          "team_id": {
            "type": "integer",
            "title": "Team Id"
          },
          "photo_url": {
            "type": "string",
            "title": "Photo Url"
          }
        },
        "type": "object",
        "required": [
          "team_id",
          "photo_url"
        ],
        "title": "SetTeamPhotoResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}